Make options required based on their type
This removes the `required` attribute option and instead determines the
requirement of a command option based on the type of the field the
`#[option]` attribute is applied on. If the type is enveloped in an
`Option<...>`, it is optional, otherwise it is required.

Additionally, the `required` attribute option has been removed from
subcommands. I do not see a point in making a subcommand mandatory,
mainly because defining subcommands disables the parent command.